+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 
 - add new config for global disable
 - add new commands to disable/enable the auto selection globally & for the opened file
+
+## 0.0.3
+
+- add a bunch of new configs
+- now we have an icon at the editor/title area to quickly disable/enable the current file selection mirroring